Fix cell set coloring
