+++ src/components/utils.js
@@ -1,5 +1,4 @@
 /* eslint-disable */
-import isEqual from 'lodash/isEqual';
 import { COORDINATE_SYSTEM } from 'deck.gl';
 
 export function makeCellStatusMessage(cellInfoFactors) {
@@ -184,29 +183,5 @@ export function mergeCellSets(cellSets, additionalCellSets) {
       ...(additionalCellSets ? additionalCellSets.tree : []),
     ],
   };
-}
-
-export function initializeCellSetColor(cellSets, cellSetColor) {
-  const nextCellSetColor = [ ...(cellSetColor || []) ];
-
-  function processNode(node, prevPath, siblingIndex) {
-    const nodePath = [...prevPath, node.name];
-    
-    const nodeColor = nextCellSetColor.find(d => isEqual(d.path, nodePath));
-
-    if(!nodeColor) {
-      nextCellSetColor.push({
-        path: nodePath,
-        color: PALETTE[siblingIndex % PALETTE.length],
-      });
-    }
-    if(node.children) {
-      node.children.forEach((c, i) => processNode(c, nodePath, i));
-    }
-  }
-
-  cellSets.tree.forEach((lzn) => processNode(lzn, [], 0));
-
-  return nextCellSetColor;
 }
